fixe  read_mobil_gate_res

+++ ClassMascaret.py
@@ -1897,7 +1897,6 @@ class ClassMascaret:
             return
         for i, scen in enumerate(dict_scen['name']):
             self.mgis.add_info("The current scenario is {}".format(scen))
-
             # initialise file
             date_debut = None
             if noyau == "steady":
@@ -2446,6 +2445,7 @@ class ClassMascaret:
                     'new simulation? '.format(nom_scen))
 
                 if ok:
+                    lst_tab = self.mdb.list_tables()
                     # delete case initalization
                     condition = "(scenario LIKE '{0}' OR  scenario " \
                                 "LIKE '{0}_init')" \
@@ -2460,9 +2460,9 @@ class ClassMascaret:
                         self.mdb.SCHEMA, run, nom_scen), fetch=True)
 
                     self.mdb.delete('runs', condition)
-                    self.mdb.delete('resultats', condition)
+                    if 'resultats' in lst_tab:
+                        self.mdb.delete('resultats', condition)
                     # new results
-
                     if len(id_run) > 0:
                         id_run = id_run[0][0]
                         condition = "id_runs = {}".format(id_run)
@@ -2478,6 +2478,8 @@ class ClassMascaret:
                             self.mdb.SCHEMA, ','.join(list_var)))
                         self.mdb.delete('results', condition)
                         self.mdb.delete('results_sect', condition)
+                        self.mdb.delete('runs_graph', condition)
+                        self.mdb.delete('runs_plani', condition)
                     if self.mgis.DEBUG:
                         self.mgis.add_info(
                             "Deletion of {0} scenario for {1} is done".format(
@@ -2583,7 +2585,7 @@ class ClassMascaret:
         """
 
         info = self.mdb.select('weirs', where="active_mob = true",
-                               list_var=['method_mob', 'gid', 'name'])
+                               list_var=['method_mob', 'gid', 'name'], order='gid')
         if info:
             try:
                 nomfich = os.path.join(self.dossierFileMasc,
@@ -2648,6 +2650,7 @@ class ClassMascaret:
                                 info['name'][i]))
 
                 fich.close()
+
                 self.mgis.add_info("Creation the dam is done")
 
             except Exception as e:
@@ -2677,49 +2680,61 @@ class ClassMascaret:
         """
         nomfich = os.path.join(self.dossierFileMasc, 'Fichier_Crete.csv')
         if os.path.isfile(nomfich):
-            # Read file
-            dico_res = {}
-            fich = open(nomfich, 'r')
+            try:
+                # Read file
+                dico_res = {}
+                fich = open(nomfich, 'r')
 
-            for ligne in fich:
-                liste = ligne.split()
-                if len(liste) > 1:
-                    nom = liste[2].strip()
-                    if not (nom in dico_res.keys()):
-                        dico_res[nom] = {'TIME': [], 'ZSTR': []}
-                    dico_res[nom]['TIME'].append(float(liste[0].strip()))
-                    dico_res[nom]['ZSTR'].append(float(liste[1].strip()))
+                for ligne in fich:
+                    liste = ligne.split()
+                    if len(liste) > 1:
+                        nom = liste[2].strip()
+                        if not (nom in dico_res.keys()):
+                            dico_res[nom] = {'TIME': [], 'ZSTR': []}
+                        dico_res[nom]['TIME'].append(float(liste[0].strip()))
+                        dico_res[nom]['ZSTR'].append(float(liste[1].strip()))
+                fich.close()
+                var_info = {'var': 'ZSTR',
+                            'type_res': 'weirs',
+                            'name': 'valve movement',
+                            'type_var': 'float'}
+                id_var = self.mdb.check_id_var(var_info)
+                # Stock information
+                colonnes = ['id_runs', 'time', 'pknum', 'var', 'val']
+                values = []
+                dico_pk = {}
+                dico_time = {}
 
-            var_info = {'var': 'ZSTR',
-                        'type_res': 'weirs',
-                        'name': 'valve movement',
-                        'type_var': 'float'}
-            id_var = self.mdb.check_id_var(var_info)
-            # Stock information
-            colonnes = ['id_runs', 'time', 'pknum', 'var', 'val']
-            values = []
-            dico_pk = {}
-            dico_time = {}
-            for name in dico_res.keys():
-                where = "name = '{}'".format(name)
-                info = self.mdb.select('weirs', where=where,
-                                       list_var=['gid', 'abscissa'])
-                time = dico_res[name]['TIME']
-                lpk = [info['abscissa'][0] for i in range(len(time))]
-                dico_pk[name] = info['abscissa'][0]
-                dico_time[name] = list(time)
-                v_tmp = self.creat_values(id_run, id_var, lpk,
-                                          time, dico_res[name]['ZSTR'])
-                values += v_tmp
-            if len(values) > 0:
-                self.mdb.insert_res('results', values, colonnes)
 
-            if len(dico_res.keys()) > 0:
-                list_insert = [[id_run, 'weirs', 'pknum', json.dumps(dico_pk)],
-                               [id_run, 'weirs', 'time', json.dumps(dico_time)],
-                               [id_run, 'weirs', 'var', json.dumps([id_var])]]
-                col_tab = ['id_runs', 'type_res', 'var', 'val']
-                self.mdb.insert_res('runs_graph', list_insert, col_tab)
+                for name in dico_res.keys():
+                    where = "name = '{}'".format(name)
+                    info = self.mdb.select('weirs', where=where,
+                                           list_var=['gid', 'abscissa'], order='gid', verbose=False)
+                    if  len(info['gid']) > 1:
+                        where = "name LIKE '{}%'".format(name)
+                        info = self.mdb.select('weirs', where=where,
+                                               list_var=['gid', 'abscissa'], order='gid', verbose=False)
+                    time = dico_res[name]['TIME']
+                    dico_time[name] = list(time)
+                    lpk = [info['abscissa'][0] for i in range(len(time))]
+                    dico_pk[name] = info['abscissa'][0]
+
+                    v_tmp = self.creat_values(id_run, id_var, lpk,
+                                              time, dico_res[name]['ZSTR'])
+                    values += v_tmp
+                if len(values) > 0:
+                    self.mdb.insert_res('results', values, colonnes)
+
+                if len(dico_res.keys()) > 0:
+                    list_insert = [[id_run, 'weirs', 'pknum', json.dumps(dico_pk)],
+                                   [id_run, 'weirs', 'time', json.dumps(dico_time)],
+                                   [id_run, 'weirs', 'var', json.dumps([id_var])]]
+                    col_tab = ['id_runs', 'type_res', 'var', 'val']
+                    self.mdb.insert_res('runs_graph', list_insert, col_tab)
+            except Exception as e:
+                txt =  "Erreur load of mobil_gate results.\n"
+                self.mgis.add_info(txt)
+                self.mgis.add_info(e)
 
     def insert_id_run(self, run, scen):
         """
@@ -3184,7 +3199,7 @@ class ClassMascaret:
                     typ_law)
             # self.mgis.add_info('{}'.format(condition))
 
-            config = self.mdb.select_one('law_config', condition, verbose=True)
+            config = self.mdb.select_one('law_config', condition, verbose=False)
 
             if config:
                 values = self.mdb.select("law_values",
